#35 Сделал карусель для отзывов и акций. В дальнейшем будет решаться как именно будут приходить данные
