+++ frontend/src/App.js
@@ -2,7 +2,6 @@ import React from "react";
 import AppToolbar from "./components/UI/Toolbar/AppToolbar";
 import Routes from "./routes/Routes";
 import {ToastContainer} from "react-toastify";
-import CarouselReviews from "./components/CarouselReviews/CarouselReviews";
 
 const App = () => {
