+++ frontend/src/containers/TestSubject/TestSubject.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Carousel from "../../components/Carousel/Carousel";
+
+const TestSubject = () => {
+    return (
+        <>
+            <Carousel type="events"/>
+            <Carousel/>
+        </>
+    );
+};
+
+export default TestSubject;
